Chore/fast coverage shards (#160)
* docs(publishing-studio): publish workflow video proof

* chore: consolidate package release work

+++ tests/Unit/WidgetExtensions/WidgetExtensionViewResolverTest.php
@@ -30,6 +30,9 @@ afterEach(function (): void {
     );
 
     foreach ($files as $file) {
+        if (! $file instanceof SplFileInfo) {
+            continue;
+        }
         $file->isDir() ? rmdir($file->getPathname()) : unlink($file->getPathname());
     }
